move routes to routes.rs, use lib.rs for app creation
